refactor hits and home screen
