+++ lib/ui/screens/hits/search_results_screen.dart
@@ -3,9 +3,9 @@ import 'package:flutter_ecom_demo/data/product_repository.dart';
 import 'package:flutter_ecom_demo/domain/product.dart';
 import 'package:flutter_ecom_demo/domain/query.dart';
 import 'package:flutter_ecom_demo/ui/product_screen.dart';
+import 'package:flutter_ecom_demo/ui/screens/hits/paged_hits_grid.dart';
+import 'package:flutter_ecom_demo/ui/screens/hits/paged_hits_list.dart';
 import 'package:flutter_ecom_demo/ui/widgets/header.dart';
-import 'package:flutter_ecom_demo/ui/widgets/product_card_view.dart';
-import 'package:flutter_ecom_demo/ui/widgets/product_item_view.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
 
 class SearchResultsScreen extends StatefulWidget {
@@ -22,8 +22,7 @@ class _SearchResultsScreen extends State<SearchResultsScreen> {
 
   Query get _query => widget.query;
   int _resultsCount = 0;
-
-  bool _isGrid = true;
+  _HitsDisplay _display = _HitsDisplay.grid;
 
   final PagingController<int, Product> _pagingController =
       PagingController(firstPageKey: 0);
@@ -96,10 +95,12 @@ class _SearchResultsScreen extends State<SearchResultsScreen> {
                     splashRadius: 10,
                     padding: const EdgeInsets.all(0.0),
                     onPressed: () => setState(() {
-                      _isGrid = true;
+                      _display = _HitsDisplay.grid;
                     }),
                     icon: Icon(Icons.grid_view,
-                        size: 20, color: _isGrid ? Colors.blue : null),
+                        size: 20,
+                        color:
+                            _HitsDisplay.grid == _display ? Colors.blue : null),
                   ),
                 ),
                 const SizedBox(width: 10),
@@ -110,10 +111,12 @@ class _SearchResultsScreen extends State<SearchResultsScreen> {
                     splashRadius: 10,
                     padding: const EdgeInsets.all(0.0),
                     onPressed: () => setState(() {
-                      _isGrid = false;
+                      _display = _HitsDisplay.list;
                     }),
                     icon: Icon(Icons.view_list,
-                        size: 20, color: !_isGrid ? Colors.blue : null),
+                        size: 20,
+                        color:
+                            _HitsDisplay.list == _display ? Colors.blue : null),
                   ),
                 ),
               ],
@@ -123,58 +126,34 @@ class _SearchResultsScreen extends State<SearchResultsScreen> {
           Expanded(
             child: Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 8),
-                child: _isGrid ? productsGrid() : productsList()),
+                child: _hitsDisplay()),
           ),
         ],
       )),
     );
   }
 
-  void presentProductPage(BuildContext context, String productID) {
+  Widget _hitsDisplay() {
+    switch (_display) {
+      case _HitsDisplay.list:
+        return PagedHitsList(
+            pagingController: _pagingController,
+            onHitClick: (objectID) => _presentProductPage(context, objectID));
+      case _HitsDisplay.grid:
+        return PagedHitsGrid(
+            pagingController: _pagingController,
+            onHitClick: (objectID) => _presentProductPage(context, objectID));
+    }
+  }
+
+  void _presentProductPage(BuildContext context, String productID) {
     _productRepository
         .getProduct(productID)
         .then((product) => Navigator.push(context, MaterialPageRoute(
-              builder: (BuildContext context) {
-                return ProductScreen(product: product);
-              },
-            )));
-  }
-
-  Widget productsGrid() {
-    return PagedGridView<int, Product>(
-      shrinkWrap: true,
-      pagingController: _pagingController,
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-        childAspectRatio: 0.9,
-        crossAxisSpacing: 20,
-        mainAxisSpacing: 20,
-        crossAxisCount: 2,
-      ),
-      builderDelegate: PagedChildBuilderDelegate<Product>(
-        itemBuilder: (context, item, index) => ProductCardView(
-            product: item,
-            imageAlignment: Alignment.bottomCenter,
-            onProductPressed: (objectID) {
-              presentProductPage(context, objectID);
-            }),
-      ),
-    );
-  }
-
-  Widget productsList() {
-    return PagedListView<int, Product>.separated(
-      shrinkWrap: true,
-      pagingController: _pagingController,
-      separatorBuilder: (context, index) => const SizedBox(height: 10),
-      builderDelegate: PagedChildBuilderDelegate<Product>(
-        itemBuilder: (context, item, index) => ProductItemView(
-            product: item,
-            imageAlignment: Alignment.bottomCenter,
-            onProductPressed: (objectID) {
-              presentProductPage(context, objectID);
-            }),
-      ),
-    );
+      builder: (BuildContext context) {
+        return ProductScreen(product: product);
+      },
+    )));
   }
 
   @override
@@ -183,3 +162,5 @@ class _SearchResultsScreen extends State<SearchResultsScreen> {
     super.dispose();
   }
 }
+
+enum _HitsDisplay { list, grid }
